Working Code
This main objective of the code is operating normally.
Some minor adjustments should be made to made the code cleaner and correct on all fronts.

+++ OneLineTrainTimeScheduling/src/Train/Position.java
@@ -17,7 +17,7 @@ public class Position {
 		this.dist = dist;
 	}
 
-	public Position(Segment seg, int dist) {
+	public Position(Segment seg, double dist) {
 		span = new Path();
 		span.addSegment(seg);
 		this.dist = dist;
